+++ tests/modeling/test_tmdd.py
@@ -294,4 +294,8 @@ def _add_random_dvids(model):
     df = model.dataset.copy()
     rng = create_rng(23)
     df['DVID'] = np.array([rng.integers(1, 6) for i in range(len(df))])
-    return model.replace(dataset=df)
+    model = model.replace(dataset=df)
+    di = model.datainfo
+    col = di['DVID'].replace(type='dvid')
+    di = di.set_column(col)
+    return model.replace(datainfo=di)